Extracting RepositoryTrackWithBenchmarking statistics section to BenchmarkedRepository interface
